Make artefact dialog a page instead
