+++ frontend/lib/ui/artefact_page/artefact_dialog_info_section.dart
